明星同款添加搜索、支付宝转账

+++ src/main/java/cn/offway/zeus/service/impl/PhRefundServiceImpl.java
@@ -104,7 +104,7 @@ public class PhRefundServiceImpl implements PhRefundService {
 
 		List<RefundGoodsDto> goodsDtos = refundDto.getGoods();
 		PhOrderInfo phOrderInfo = phOrderInfoService.findByOrderNo(orderNo);
-		if(phOrderInfo.getUserId() != refundDto.getUserId()){
+		if(phOrderInfo.getUserId().longValue() != refundDto.getUserId().longValue()){
 			return jsonResultHelper.buildFailJsonResult(CommonResultCode.PARAM_ERROR);
 		}
